pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "sbmlpbkutils"
description = "Utils for FAIR PBK model development with SBML"
dynamic = ["version"]
readme = "README.md"
authors = [
    {name = "Johannes Kruisselbrink", email = "jwkruisselbrink@gmail.com"}
]
maintainers = [ 
    {name = "Johannes Kruisselbrink", email = "jwkruisselbrink@gmail.com"}
]
dependencies = [
    "PyYAML==6.0.1",
    "antimony>=2.15.0",
    "numpy>=1.26.4",
    "python-libsbml>=5.20.2",
    "sbmlutils>=0.9.0",
    "tellurium>=2.2.10",
    "owlready2>=0.47",
    "pyyaml>=6.0.1"
]
keywords=["pbk","sbml","fair"]
classifiers = [
    "Development Status :: 2 - Pre-Alpha", 
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
]
license = {file = "LICENSE"}
requires-python = ">=3.12"

[project.urls]
Homepage = "https://github.com/jwkruisselbrink/sbml-pbk-utils"
Repository = "https://github.com/jwkruisselbrink/sbml-pbk-utils"

[tool.setuptools_scm]
version_file = "sbmlpbkutils/_version.py"
